# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Brave Accounts Service",
        "contact": {}
    },
    "paths": {
...
        "/v2/verify/complete": {
            "post": {
                "description": "Completes the email verification process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email verification"
                ],
                "summary": "Complete email verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brave services key (if one is configured)",
                        "name": "Brave-Key",
                        "in": "header"
                    },
                    {
                        "description": "Verify completion params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.VerifyCompleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.VerifyCompleteResponse"
                        }
                    },
                    "400": {
                        "description": "Missing/invalid verification parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Verification not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v2/verify/email_viewer": {
            "get": {
                "description": "Retrieves and displays emails sent through LocalStack SES endpoint",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Debugging"
                ],
                "summary": "View sent emails in LocalStack SES",
                "responses": {
                    "200": {
                        "description": "HTML page displaying emails",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v2/verify/init": {
            "post": {
                "description": "Starts email verification process by sending a verification email\nOne of the following intents must be provided with the request:\n- `auth_token`: After verification, create an account if one does not exist, and generate an auth token. The token will be available via the \"query result\" endpoint.\n- `verification`: After verification, do not create an account, but indicate that the email was verified in the \"query result\" response. Do not allow registration after verification.\n- `registration`: After verification, indicate that the email was verified in the \"query result\" response. An account may be created by setting a password.\n- `set_password`: After verification, indicate that the email was verified in the \"query result\" response. A password may be set for the existing account.\n\nOne of the following service names must be provided with the request: `inbox-aliases`, `accounts`, `premium`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email verification"
                ],
                "summary": "Initialize email verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brave services key (if one is configured)",
                        "name": "Brave-Key",
                        "in": "header"
                    },
                    {
                        "description": "Verification request params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.VerifyInitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.VerifyInitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/verify/result": {
            "post": {
                "description": "Provides the status of a pending or successful verification.\nIf the wait option is set to true, the server will up to 20 seconds for verification. Feel free\nto call this endpoint repeatedly to wait for verification.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email verification"
                ],
                "summary": "Query result of verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + verification token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Brave services key (if one is configured)",
                        "name": "Brave-Key",
                        "in": "header"
                    },
                    {
                        "description": "Auth token request params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.VerifyResultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.VerifyResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.VerifyCompleteRequest": {
            "description": "Request parameters for verification completion",
            "type": "object",
            "required": [
                "code",
                "id"
            ],
            "properties": {
                "code": {
                    "description": "Verification code sent to user",
                    "type": "string"
                },
                "id": {
                    "description": "Unique verification identifier",
                    "type": "string"
                }
            }
        },
        "controllers.VerifyCompleteResponse": {
            "description": "Response for verification completion",
            "type": "object",
            "properties": {
                "service": {
                    "description": "Name of service requesting verification",
                    "type": "string"
                },
                "verificationToken": {
                    "description": "JWT token for checking verification status",
                    "type": "string"
                }
            }
        },
        "controllers.VerifyInitRequest": {
            "description": "Request to initialize email verification",
            "type": "object",
            "required": [
                "email",
                "intent",
                "service"
            ],
            "properties": {
                "email": {
                    "description": "Email address to verify",
                    "type": "string",
                    "example": "test@example.com"
                },
                "intent": {
                    "description": "Purpose of verification (e.g., get auth token, simple verification, registration)",
                    "type": "string",
                    "enum": [
                        "auth_token",
                        "verification",
                        "registration",
                        "set_password"
                    ],
                    "example": "registration"
                },
                "language": {
                    "description": "Locale for verification email",
                    "type": "string",
                    "maxLength": 8,
                    "example": "en-US"
                },
                "service": {
                    "description": "Service requesting the verification",
                    "type": "string",
                    "enum": [
                        "accounts",
                        "premium",
                        "inbox-aliases"
                    ],
                    "example": "accounts"
                }
            }
        },
        "controllers.VerifyInitResponse": {
            "description": "Response containing verification check token",
            "type": "object",
            "properties": {
                "verificationToken": {
                    "description": "JWT token for checking verification status",
                    "type": "string"
                }
            }
        },
        "controllers.VerifyResultRequest": {
            "description": "Request for getting auth token after verification",
            "type": "object",
            "properties": {
                "wait": {
                    "description": "Whether to wait for verification to complete",
                    "type": "boolean"
                }
            }
        },
        "controllers.VerifyResultResponse": {
            "description": "Response containing auth token",
            "type": "object",
            "properties": {
                "authToken": {
                    "description": "JWT auth token, null if verification incomplete or if password setup is required",
                    "type": "string"
                },
                "email": {
                    "description": "Email associated wiith the verification",
                    "type": "string"
                },
                "service": {
                    "description": "Name of service requesting verification",
                    "type": "string"
                },
                "verified": {
                    "description": "Email verification status",
                    "type": "boolean"
                }
            }
        },
...
        "util.ErrorResponse": {
            "description": "Standard error response",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        }
    },
...
}